+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.6.0.0
 * Added interactivity to the legend items
+* Added default sorting
+* Changed colors assignment
 * API 5.11.0
 
 ## 2.5.1.0