+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.6.0.0
 * Added interactivity to the legend items
 * Added default sorting
+* Added report page tooltip
+* Added sorting
 * Changed colors assignment
 * API 5.11.0